requirements.txt
certifi==2021.5.30
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.0
django-rest-coreapi-schema==0.1.2
djangorestframework==3.9.4
idna==2.10
itypes==1.2.0
Jinja2==2.11.3
Markdown==2.6.10
MarkupSafe==1.1.1
mysqlclient==2.0.3
Pygments==2.2.0
python-decouple==3.4
pytz==2021.1
requests==2.25.1
sqlparse==0.4.2
uritemplate==3.0.1
urllib3==1.26.7
